chore: Move header hero to the correct location
